remove semicolons and fix distributed utils logic

+++ src/utils/distributed.py
@@ -1,13 +1,13 @@
 # This file has been created by Ashley and Luna. It provides utility functions for distributing the program with SLURM
 
-import os;
-import time;
-from pathlib import Path;
-import shutil;
-import numpy as np;
-import utils.logging;
-import logging;
-from utils.paths import FLAGS_PARENT as FP;
+import os
+import time
+from pathlib import Path
+import shutil
+import numpy as np
+import utils.logging
+import logging
+from utils.paths import FLAGS_PARENT as FP
 
 # Source - https://stackoverflow.com/a/2785908
 # Posted by Alex Martelli, modified by community. See post 'Timeline' for change history
@@ -15,9 +15,9 @@ from utils.paths import FLAGS_PARENT as FP;
 def wait_until( somepredicate, timeout: int | None, logger, waiting_on_array: int, taskname: str, period=1, *args, **kwargs ):
     mustend = time.time() + timeout if timeout is not None else None
     while ( mustend is None ) or ( time.time() < mustend ):
-        logger.debug( f'Waiting on array {waiting_on_array} to complete {taskname}: time={time.time()}' );
+        logger.debug( f'Waiting on array {waiting_on_array} to complete {taskname}: time={time.time()}' )
         if somepredicate( *args, **kwargs ): 
-            logger.debug( f'Array {waiting_on_array} has completed {taskname}: time={time.time()}' );
+            logger.debug( f'Array {waiting_on_array} has completed {taskname}: time={time.time()}' )
             return True
         time.sleep(period)
     return False
@@ -37,37 +37,41 @@ class DistributedUtils:
     Utility functions for running on a distributed system (SLURM in particular)
     """
     def __init__( self ):
-        FP.mkdir( exist_ok=True ); #Would ideally run files.paths on a single node but that requires the FP directory be made
-        self.logger = utils.logging.get_logger( __name__, logging.DEBUG );
+        FP.mkdir( exist_ok=True ) #Would ideally run files.paths on a single node but that requires the FP directory be made
+        self.logger = utils.logging.get_logger( __name__, logging.DEBUG )
 
     def is_distributed( self ) -> bool:
-        return self.get_task_count() != 1;
+        return self.get_task_count() != 1
 
     def get_task_id( self ) -> int:
-        return int( os.environ.get( "SLURM_ARRAY_TASK_ID", 0 ) );
+        return int( os.environ.get( "SLURM_ARRAY_TASK_ID", 0 ) )
 
     def get_task_count( self ) -> int:
-        return int( os.environ.get( "SLURM_ARRAY_TASK_COUNT", 1 ) );
+        return int( os.environ.get( "SLURM_ARRAY_TASK_COUNT", 1 ) )
 
     def get_bin_end( self, n: int ) -> int:
         """
         Function to take a total number n and get the corresponding end of the bin that this node should be dealing with
         """
-        return ( n * ( self.get_task_id() + 1 ) ) // self.get_task_count();
+        return ( n * ( self.get_task_id() + 1 ) ) // self.get_task_count()
 
     def get_bin_start( self, n: int ) -> int:
         """
         Function to take a total number n and get the corresponding start of the bin that this node should be dealing with
         """
-        return ( n * self.get_task_id() ) // self.get_task_count();
+        return ( n * self.get_task_id() ) // self.get_task_count()
 
-    def single_task_only_forcewait( self, taskname: str, function, do_on_array_id: int, *args, **kwargs ):
+    def is_in_single_task_function( self ):
+        return ( os.environ.get( 'SINGLE_TASK_ONLY_FIRST', -1 ) != -1 ) or ( os.environ.get( 'SINGLE_TASK_ONLY_LAST', -1 ) != -1 )
+
+    def single_task_only_first( self, taskname: str, function, do_on_array_id: int, *args, **kwargs ):
         """
-        A method to execute a function only once when run on a job array, and force all other tasks to wait for completion.
+        A method to execute a function only once when run on a job array, and force all OTHER tasks to wait on THIS one to complete.
 
-        Wrapping everything you want to do in a single single_task_only_forcewait is sufficient to ensure the tasks
-        are performed on only one array element. It is advisable NOT to wrap single_task_only_forcewait calls inside
-        other single_task_only_forcewait calls as it will only produce unneccesary overhead
+        Wrapping everything you want to do in a single single_task_only_* is sufficient to ensure the tasks
+        are performed on only one array element. It is advisable NOT to wrap single_task_only_* calls inside
+        other single_task_only_* calls as it will only produce unneccesary overhead, and in the worst case of
+        mixing the two could result in the program hanging indefinitely
 
         Paramters
         ---------
@@ -82,55 +86,70 @@ class DistributedUtils:
             arguments to pass to the function call
         """
         if do_on_array_id >= self.get_task_count():
-            raise ValueError( f'Array id {do_on_array_id} out of range for array of length {self.get_task_count()}' );
+            raise ValueError( f'Array id {do_on_array_id} out of range for array of length {self.get_task_count()}' )
+
+        # Use environment variables to determine if we are in a single_task_only_* block for this task to prevent infinite hangs due to recursion
+        stof_block = os.environ.get( 'SINGLE_TASK_ONLY_FIRST' )
+        stol_block = os.environ.get( 'SINGLE_TASK_ONLY_LAST' )
+        if stol_block is not None and ( int( stol_block ) == do_on_array_id ):
+            raise RecursionError( f'Cannot call single_task_only_first for array {do_on_array_id} when running single_task_only_last for array {stol_block} - must not be the same task id' )
+        if ( stof_block is not None ) and ( int( stof_block ) != do_on_array_id ):
+            raise RecursionError( f'Cannot call single_task_only_first for array {do_on_array_id} when running single_task_only_first for array {stof_block} - must be the same task id' )
+        
+        os.environ[ 'SINGLE_TASK_ONLY_FIRST' ] = str( do_on_array_id )
+        
     
-        taskname = taskname.replace( '/', '_' ).replace( '-', '_' ).upper();
+        taskname = taskname.replace( '/', '_' ).replace( '-', '_' ).upper()
 
         if not self.is_distributed():
             # Doing this separately from if we have multiple array elements allows us to ignore flag file
             # deletion when this array id == do_on_array_id - hereafter we can assume if self.get_task_id() == do_on_array_id
             # that it is the first array to attempt the problem
-            function( *args, **kwargs );
+            function( *args, **kwargs )
         else:
             if self.get_task_id() == do_on_array_id:
-                function( *args, **kwargs );
-                ( FP / f'TASK_{taskname}_COMPLETED' ).touch();
+                function( *args, **kwargs )
+                ( FP / f'TASK_{taskname}_COMPLETED' ).touch()
             else:
                 # Truth value tells us whether or not wait_until timed out
                 def __is_completed_lambda( taskname ):
-                    return ( FP / f'TASK_{taskname}_COMPLETED' ).exists();
+                    return ( FP / f'TASK_{taskname}_COMPLETED' ).exists()
                 if not wait_until( __is_completed_lambda, None, self.logger, do_on_array_id, taskname, 1, taskname ):
-                    raise RuntimeError( f'ERROR - wait_until timed out on array {self.get_task_id()} waiting for {taskname} on array {do_on_array_id}' );
+                    raise RuntimeError( f'ERROR - wait_until timed out on array {self.get_task_id()} waiting for {taskname} on array {do_on_array_id}' )
 
-                ( FP / f'TASK_{taskname}_ARRAY_{self.get_task_id()}_PASS_COMPLETE' ).touch();
+                ( FP / f'TASK_{taskname}_ARRAY_{self.get_task_id()}_PASS_COMPLETE' ).touch()
 
                 # If we're the last array to pass this function, delete the environment variables behind us
                 # also we don't need to check for the completed variable since it must exist to get here
-                last_array = True;
+                last_array = True
                 for i in range( 0, self.get_task_count() ):
                     if i == do_on_array_id:
-                        continue;
+                        continue
                     elif not ( FP / f'TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE' ).exists():
-                        last_array = False;
+                        last_array = False
                 if last_array:
                     try:
-                        ( FP / f'TASK_{taskname}_COMPLETED' ).unlink();
+                        ( FP / f'TASK_{taskname}_COMPLETED' ).unlink()
                     except FileNotFoundError:
-                        self.logger.warning( f'Tried to delete TASK_{taskname}_ARRAY simultaneously with another node. This is probably harmless.' );
+                        self.logger.warning( f'Tried to delete TASK_{taskname}_ARRAY simultaneously with another node. This is probably harmless.' )
                     for i in range( 0, self.get_task_count() ):
                         if i == do_on_array_id:
-                            continue;
+                            continue
                         else: 
                             try:
-                                ( FP / f'TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE' ).unlink();
+                                ( FP / f'TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE' ).unlink()
                             except FileNotFoundError:
-                                self.logger.warning( f'Tried to delete TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE simultaneously with another node. This is probably harmless.' );
+                                self.logger.warning( f'Tried to delete TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE simultaneously with another node. This is probably harmless.' )
+        os.environ.pop( 'SINGLE_TASK_ONLY_FIRST' )
     
-    def last_task_only( self, taskname: str, function, *args, **kwargs ):
+    def single_task_only_last( self, taskname: str, function, do_on_array_id: int, *args, **kwargs ):
         """
-        Similarly to single_task_only_forcewait this method is used to run a function only once on a single node,
-        however this method executes the task on the last node to run. This is useful e.g. if all the tasks are
-        doing data preparation and something is to be done once the data is prepared, like graphing the data.
+        A method to execute a function only once when run on a job array, and force THIS task to wait on OTHERS to complete.
+
+        Wrapping everything you want to do in a single single_task_only_* is sufficient to ensure the tasks
+        are performed on only one array element. It is advisable NOT to wrap single_task_only_* calls inside
+        other single_task_only_* calls as it will only produce unneccesary overhead, and in the worst case of
+        mixing the two could result in the program hanging indefinitely
 
         Paramters
         ---------
@@ -138,29 +157,57 @@ class DistributedUtils:
             The name of the task - should be unique and be a valid str to include in a path name
         function
             The function to call only once.
+        do_on_array_id : int
+            The array id on which to call the function. All other array ids must execute before this function
+ 
+        *args, **kwargs
+            arguments to pass to the function call
         """
-        taskname = taskname.replace( '/', '_' ).replace( '-', '_' ).upper();
+        if do_on_array_id >= self.get_task_count():
+            raise ValueError( f'Array id {do_on_array_id} out of range for array of length {self.get_task_count()}' )
+
+        # Use environment variables to determine if we are in a single_task_only_* block for this task to prevent infinite hangs due to recursion
+        stof_block = os.environ.get( 'SINGLE_TASK_ONLY_FIRST' )
+        stol_block = os.environ.get( 'SINGLE_TASK_ONLY_LAST' )
+        if stol_block is not None and ( int( stol_block ) == do_on_array_id ):
+            raise RecursionError( f'Cannot call single_task_only_last for array {do_on_array_id} when running single_task_only_last for array {stol_block} - must be the same task id' )
+        if ( stof_block is not None ):
+            raise RecursionError( f'Cannot call single_task_only_last for array {do_on_array_id} when running single_task_only_first for array {stof_block} - fundamental recursion error' )
+        
+        os.environ[ 'SINGLE_TASK_ONLY_LAST' ] = str( do_on_array_id )
+    
+        taskname = taskname.replace( '/', '_' ).replace( '-', '_' ).upper()
 
         if not self.is_distributed():
-            # Doing this separately from if we have multiple array elements allows us to ignore flag files
-            function( *args, **kwargs );
+            # Doing this separately from if we have multiple array elements allows us to ignore flag file
+            # deletion when this array id == do_on_array_id - hereafter we can assume if self.get_task_id() == do_on_array_id
+            # that it is the first array to attempt the problem
+            function( *args, **kwargs )
         else:
-            last_array = True;
-            for i in range( 0, self.get_task_count() ):
-                if i == self.get_task_id():
-                    continue;
-                elif not ( FP / f'TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE' ).exists():
-                    last_array = False;
-
-            if last_array:
-                for i in range( 0, self.get_task_count() ):
-                    if i == self.get_task_id():
-                        continue;
-                    else: ( FP / f'TASK_{taskname}_ARRAY_{i}_PASS_COMPLETE' ).unlink();
-                function( *args, **kwargs );
+            if self.get_task_id() != do_on_array_id:
+                ( FP / f'ARRAY_{self.get_task_id()}_TASK_{taskname}_PASS' ).touch()
             else:
-                ( FP / f'TASK_{taskname}_ARRAY_{self.get_task_id()}_PASS_COMPLETE' ).touch();
+                # Truth value tells us whether or not wait_until timed out
+                def __other_arrays_passed_lambda( taskname ):
+                    other_arrays_passed = True
+                    for i in range( 0, self.get_task_count() ):
+                        if i == self.get_task_id():
+                            continue
+                        elif not ( FP / f'ARRAY_{i}_TASK_{taskname}_PASS' ).exists():
+                            other_arrays_passed = False
+                    return other_arrays_passed
 
+                if not wait_until( __other_arrays_passed_lambda, None, self.logger, do_on_array_id, taskname, 1, taskname ):
+                    raise RuntimeError( f'ERROR - wait_until timed out on array {self.get_task_id()} waiting for {taskname} on array {do_on_array_id}' )
+
+                function( *args, **kwargs )
+                for i in range( 0, self.get_task_count() ):
+                    if i == do_on_array_id:
+                        continue
+                    else: 
+                        ( FP / f'ARRAY_{i}_TASK_{taskname}_PASS' ).unlink()
+        os.environ.pop( 'SINGLE_TASK_ONLY_LAST' )
+    
 
     def copy_file_for_multiple_nodes( self, file: Path ):
         """
@@ -175,7 +222,7 @@ class DistributedUtils:
         def __copy_file_lambda():
             for i in range( self.get_task_count() ):
                 shutil.copyfile( str(file), f'{str(file.parent)}/{file.stem}_{i}{file.suffix}' )
-        self.single_task_only_forcewait( f'copy_file_{str(file)}', __copy_file_lambda, 0 );
+        self.single_task_only_first( f'copy_file_{str(file)}', __copy_file_lambda, 0 )
